Tests for User Resources (#2035)
* Test the LinuxVM

* Thinned out helpers.py

* Split register bundles out

* Putting build back

* Grouped workspace helper functions

* Reuse GetAuthHeaders

* Split publish bundles out
